requirements.txt
Flask==2.3.2
protobuf==5.28.3
PyPDF2==3.0.1
google-generativeai